docs: updating Image docs
