+++ docs/src/pages/components/image/examples/DefaultImageExample.tsx
@@ -1,0 +1,12 @@
+import { Image } from '@aws-amplify/ui-react';
+
+export const DefaultImageExample = () => {
+  return (
+    <Image
+      src="/road-to-milford-new-zealand-800w.jpg"
+      alt="View from road to Milford Sound, New Zealand.
+  Glittering stream with old log, snowy mountain peaks
+  tower over a green field."
+    />
+  );
+};
