+++ docs/src/pages/components/image/examples/index.ts
@@ -1,0 +1,3 @@
+export { DefaultImageExample } from './DefaultImageExample';
+export { ImageObjectFitAndPositionExample } from './ImageObjectFitAndPositionExample';
+export { ImageOptimizationExample } from './ImageOptimizationExample';
