+++ docs/src/pages/components/image/index.page.mdx
@@ -1,5 +1,6 @@
 ---
 title: Image
+description: The Image primitive can be used to display responsive images.
 isPrimitive: true
 ---
