+++ docs/src/pages/components/image/react.mdx
@@ -1,7 +1,12 @@
 import { Image, View } from '@aws-amplify/ui-react';
-import { ImageDemo } from './demo';
 
-The Image primitive can be used to display responsive images.
+import { Example, ExampleCode } from '@/components/Example';
+import { ImageDemo } from './demo';
+import {
+  DefaultImageExample,
+  ImageObjectFitAndPositionExample,
+  ImageOptimizationExample,
+} from './examples';
 
 ## Demo
 
@@ -9,76 +14,42 @@ The Image primitive can be used to display responsive images.
 
 ## Usage
 
-Import the `Image` component and styles. Images are responsive by default (if you're on desktop, try resizing your browser window).
+Import the `Image` component. Images are responsive by default (if you're on desktop, try resizing your browser window).
 
-```jsx
-<Image
-  src="/road-to-milford-new-zealand-800w.jpg"
-  alt="View from road to Milford Sound, New Zealand.
-  Glittering stream with old log, snowy mountain peaks
-  tower over a green field."
-/>
-```
+<Example>
+  <DefaultImageExample />
+  <ExampleCode>
+    ```jsx file=./examples/DefaultImageExample.tsx
 
-<Image
-  src="/road-to-milford-new-zealand-800w.jpg"
-  alt="View from road to Milford Sound, New Zealand.
-  Glittering stream with old log, snowy mountain peaks
-  tower over a green field."
-/>
+    ```
+
+  </ExampleCode>
+</Example>
 
 ### Responsive Image optimization
 
 Use `sizes` and `srcSet` to dynamically load different Image sizes based on screen size/resolution. For more information, see [MDN responsive images article](https://developer.mozilla.org/en-US/docs/Learn/HTML/Multimedia_and_embedding/Responsive_images).
 
-```jsx
-<Image
-  srcSet="/road-to-milford-new-zealand-800w.jpg 800w,
-          /road-to-milford-new-zealand-1200w.jpg 1200w,
-          /road-to-milford-new-zealand-1400w.jpg 1400w"
-  alt="View from road to Milford Sound, New Zealand.
-      Glittering stream with old log, snowy mountain peaks
-      tower over a green field."
-/>
-```
+<Example>
+  <ImageOptimizationExample />
+  <ExampleCode>
+    ```jsx file=./examples/ImageOptimizationExample.tsx
 
-<Image
-  srcSet="/road-to-milford-new-zealand-800w.jpg 800w,
-          /road-to-milford-new-zealand-1200w.jpg 1200w,
-          /road-to-milford-new-zealand-1400w.jpg 1400w"
-  alt="View from road to Milford Sound, New Zealand.
-      Glittering stream with old log, snowy mountain peaks
-      tower over a green field."
-/>
+    ```
+
+  </ExampleCode>
+</Example>
 
 ### Object fit and object position
 
 To control how an Image fits its container, use the `objectFit` and `objectPosition` properties.
 
-```jsx
-<View width="400px" height="400px" boxShadow="8px 6px 6px 0 #ccc">
-  <Image
-    width="100%"
-    height="100%"
-    objectFit="cover"
-    src="/road-to-milford-new-zealand-800w.jpg"
-    alt="Glittering stream with old log, snowy mountain peaks
-    tower over a green field."
-  />
-  <p>Road to Milford Sound, New Zealand. Photo by Scott Rees.</p>
-</View>
-```
+<Example>
+  <ImageObjectFitAndPositionExample />
+  <ExampleCode>
+    ```jsx file=./examples/ImageObjectFitAndPositionExample.tsx
 
-<View width="400px" height="400px" boxShadow="8px 6px 6px 0 #ccc">
-  <Image
-    width="100%"
-    height="100%"
-    objectFit="cover"
-    src="/road-to-milford-new-zealand-800w.jpg"
-    alt="Glittering stream with old log, snowy mountain peaks
-    tower over a green field."
-  />
-  <p className="text-sm m-0 p-0">
-    Road to Milford Sound, New Zealand. Photo by Scott Rees.
-  </p>
-</View>
+    ```
+
+  </ExampleCode>
+</Example>
