+++ docs/src/pages/components/shared/useStyleProps.tsx
@@ -1,5 +1,6 @@
-import { useState, useEffect } from 'react';
+import * as React from 'react';
 import { BaseStyleProps } from '@aws-amplify/ui-react';
+
 import { StylePropControlsProps } from './StylePropControls';
 
 interface UseStyleProps {
@@ -7,43 +8,43 @@ interface UseStyleProps {
 }
 
 export const useStyleProps: UseStyleProps = (initialValues) => {
-  const [width, setWidth] = useState<BaseStyleProps['width']>(
+  const [width, setWidth] = React.useState<BaseStyleProps['width']>(
     initialValues.width
   );
-  const [height, setHeight] = useState<BaseStyleProps['height']>(
+  const [height, setHeight] = React.useState<BaseStyleProps['height']>(
     initialValues.height
   );
-  const [maxHeight, setMaxHeight] = useState<BaseStyleProps['maxHeight']>(
+  const [maxHeight, setMaxHeight] = React.useState<BaseStyleProps['maxHeight']>(
     initialValues.maxHeight
   );
-  const [minHeight, setMinHeight] = useState<BaseStyleProps['minHeight']>(
+  const [minHeight, setMinHeight] = React.useState<BaseStyleProps['minHeight']>(
     initialValues.minHeight
   );
-  const [maxWidth, setMaxWidth] = useState<BaseStyleProps['maxWidth']>(
+  const [maxWidth, setMaxWidth] = React.useState<BaseStyleProps['maxWidth']>(
     initialValues.maxWidth
   );
-  const [minWidth, setMinWidth] = useState<BaseStyleProps['minWidth']>(
+  const [minWidth, setMinWidth] = React.useState<BaseStyleProps['minWidth']>(
     initialValues.minWidth
   );
-  const [color, setColor] = useState<BaseStyleProps['color']>(
+  const [color, setColor] = React.useState<BaseStyleProps['color']>(
     initialValues.color
   );
-  const [backgroundColor, setBackgroundColor] = useState<
+  const [backgroundColor, setBackgroundColor] = React.useState<
     BaseStyleProps['backgroundColor']
   >(initialValues.backgroundColor);
-  const [boxShadow, setBoxShadow] = useState<BaseStyleProps['boxShadow']>(
+  const [boxShadow, setBoxShadow] = React.useState<BaseStyleProps['boxShadow']>(
     initialValues.boxShadow
   );
-  const [padding, setPadding] = useState<BaseStyleProps['padding']>(
+  const [padding, setPadding] = React.useState<BaseStyleProps['padding']>(
     initialValues.padding
   );
-  const [border, setBorder] = useState<BaseStyleProps['border']>(
+  const [border, setBorder] = React.useState<BaseStyleProps['border']>(
     initialValues.border
   );
-  const [borderRadius, setBorderRadius] = useState<
+  const [borderRadius, setBorderRadius] = React.useState<
     BaseStyleProps['borderRadius']
   >(initialValues.borderRadius);
-  const [opacity, setOpacity] = useState<BaseStyleProps['opacity']>(
+  const [opacity, setOpacity] = React.useState<BaseStyleProps['opacity']>(
     initialValues.opacity
   );
